Removed Shared Library Nonsense. Finished Assignment09

+++ Assignment08/test08.c
@@ -12,25 +12,25 @@ void testMatrixKronecker() {
     int s2 = 7;
     int s3 = 3;
     int s4 = 4;
-    float matrix1Template[5][7] = 
+    double matrix1Template[5][7] = 
         {{4, 4, 6, 1, 7, 3, 7},
         {1, 6, 0, 5, 6, 0, 0},
         {6, 7, 6, 2, 9, 3, 9},
         {6, 7, 6, 9, 8, 6, 5},
         {4, 4, 3, 0, 6, 4, 8}};
-    float matrix2Template[3][4] = 
+    double matrix2Template[3][4] = 
         {{4, 3, 4, 4},
         {1, 2, 1, 7},
         {9, 3, 7, 0}};
 
-    float** matrix1 = (float**)malloc(s1 * sizeof(float*));
+    double** matrix1 = (double**)malloc(s1 * sizeof(double*));
     for (int i = 0; i < s1; i++) {
-        matrix1[i] = (float*)malloc(s2 * sizeof(float));
+        matrix1[i] = (double*)malloc(s2 * sizeof(double));
     }
 
-    float** matrix2 = (float**)malloc(s3 * sizeof(float*));
+    double** matrix2 = (double**)malloc(s3 * sizeof(double*));
     for (int i = 0; i < s3; i++) {
-        matrix2[i] = (float*)malloc(s4 * sizeof(float));
+        matrix2[i] = (double*)malloc(s4 * sizeof(double));
     }
 
     for (int i = 0; i < s1; i++) {
@@ -46,7 +46,7 @@ void testMatrixKronecker() {
     }
 
     double time;
-    float** result = matrixKronecker(matrix1, matrix2, s1, s2, s3, s4, &time);
+    double** result = matrixKronecker(matrix1, matrix2, s1, s2, s3, s4, &time);
 
     printf("Result:\n");
     for (int i = 0; i < s1 * s3; i++) {
@@ -69,14 +69,14 @@ void timeMatrixKronecker() {
         int s1 = k;
         int s2 = k;
 
-        float** matrix1 = (float**)malloc(s1 * sizeof(float*));
+        double** matrix1 = (double**)malloc(s1 * sizeof(double*));
         for (int i = 0; i < s1; i++) {
-            matrix1[i] = (float*)malloc(s2 * sizeof(float));
+            matrix1[i] = (double*)malloc(s2 * sizeof(double));
         }
 
-        float** matrix2 = (float**)malloc(s1 * sizeof(float*));
+        double** matrix2 = (double**)malloc(s1 * sizeof(double*));
         for (int i = 0; i < s2; i++) {
-            matrix2[i] = (float*)malloc(s2 * sizeof(float));
+            matrix2[i] = (double*)malloc(s2 * sizeof(double));
         }
 
         for (int i = 0; i < s1; i++) {
@@ -87,27 +87,27 @@ void timeMatrixKronecker() {
         }
 
         double time;
-        float** result = matrixKronecker(matrix1, matrix2, s1, s2, s1, s2, &time);
+        double** result = matrixKronecker(matrix1, matrix2, s1, s2, s1, s2, &time);
 
         printf("Size %i: (%f)\n", k, time);
     }
 }
 
-void testPower1(float (*f)()) {
+void testPower1(double (*f)()) {
     int s1 = 5;
     int s2 = 5;
-    float matrix1Template[5][5] = 
+    double matrix1Template[5][5] = 
                     {{8, 4, 3, 4, 6},
                     {9, 5, 5, 8, 3},
                     {1, 3, 3, 4, 1},
                     {8, 9, 7, 0, 1},
                     {0, 3, 3, 3, 3}};
     
-    float * vector = (float*)malloc(s1 * sizeof(float));
+    double * vector = (double*)malloc(s1 * sizeof(double));
 
-    float** matrix1 = (float**)malloc(s1 * sizeof(float*));
+    double** matrix1 = (double**)malloc(s1 * sizeof(double*));
     for (int i = 0; i < s1; i++) {
-        matrix1[i] = (float*)malloc(s2 * sizeof(float));
+        matrix1[i] = (double*)malloc(s2 * sizeof(double));
     }
 
     for (int i = 0; i < s1; i++) {
@@ -118,17 +118,17 @@ void testPower1(float (*f)()) {
     }
 
     double time;
-    float result = f(matrix1, vector, 0.00001, 1000, s1, &time);
+    double result = f(matrix1, vector, 0.00001, 1000, s1, &time);
 
     printf("Result: %f\n", result);
     printf("In %f Seconds\n", time);
 
 }
 
-void testPower2(float (*f)()) {
+void testPower2(double (*f)()) {
     int s1 = 12;
     int s2 = 12;
-    float matrix1Template[12][12] = 
+    double matrix1Template[12][12] = 
                     {{8, 4, 5, 9, 9, 8, 4, 3, 9, 6, 6, 7},
                     {0, 6, 5, 2, 9, 6, 3, 1, 8, 6, 4, 1},
                     {8, 9, 9, 5, 6, 7, 3, 5, 6, 0, 3, 9},
@@ -143,11 +143,11 @@ void testPower2(float (*f)()) {
                     {1, 6, 7, 8, 3, 4, 5, 5, 0, 3, 1, 0}}
 ;
     
-    float * vector = (float*)malloc(s1 * sizeof(float));
+    double * vector = (double*)malloc(s1 * sizeof(double));
 
-    float** matrix1 = (float**)malloc(s1 * sizeof(float*));
+    double** matrix1 = (double**)malloc(s1 * sizeof(double*));
     for (int i = 0; i < s1; i++) {
-        matrix1[i] = (float*)malloc(s2 * sizeof(float));
+        matrix1[i] = (double*)malloc(s2 * sizeof(double));
     }
 
     for (int i = 0; i < s1; i++) {
@@ -158,7 +158,7 @@ void testPower2(float (*f)()) {
     }
 
     double time;
-    float result = f(matrix1, vector, 0.00001, 1000, s1, &time);
+    double result = f(matrix1, vector, 0.00001, 1000, s1, &time);
 
     printf("Result: %f\n", result);
     printf("In %f Seconds\n", time);
@@ -172,10 +172,10 @@ void timeFunctions() {
         int s1 = k;
         int s2 = k;
 
-        float** matrix1 = (float**)malloc(s1 * sizeof(float*));
-        float * vector = (float*)malloc(s1 * sizeof(float));
+        double** matrix1 = (double**)malloc(s1 * sizeof(double*));
+        double * vector = (double*)malloc(s1 * sizeof(double));
         for (int i = 0; i < s1; i++) {
-            matrix1[i] = (float*)malloc(s2 * sizeof(float));
+            matrix1[i] = (double*)malloc(s2 * sizeof(double));
         }
 
         for (int i = 0; i < s1; i++) {
@@ -188,9 +188,9 @@ void timeFunctions() {
         double time1;
         double time2;
         double time3;
-        float result1 = powerSimple(matrix1, vector, 0.0001, 1000, s1, &time1);
-        float result2 = powerOptimized(matrix1, vector, 0.0001, 1000, s1, &time2);
-        float result3 = powerOMP(matrix1, vector, 0.0001, 1000, s1, &time3);
+        double result1 = powerSimple(matrix1, vector, 0.0001, 1000, s1, &time1);
+        double result2 = powerOptimized(matrix1, vector, 0.0001, 1000, s1, &time2);
+        double result3 = powerOMP(matrix1, vector, 0.0001, 1000, s1, &time3);
 
         printf("Size %i\n", k);
         printf("SIMPLE: %f\n", time1);
@@ -204,18 +204,18 @@ void testJacoby() {
     int s1 = 3;
     int s2 = 3;
 
-    float matrix1Template[3][3] = 
+    double matrix1Template[3][3] = 
         {{3, -2, 1},
         {1, -3, 2},
         {-1, 2, 4}};
 
-    float vector1Template[4] = {2, 1, 3};
+    double vector1Template[4] = {2, 1, 3};
 
-    float * vector1 = (float*)malloc(s1 * sizeof(float));
-    float * vector2 = (float*)malloc(s1 * sizeof(float));
-    float** matrix1 = (float**)malloc(s1 * sizeof(float*));
+    double * vector1 = (double*)malloc(s1 * sizeof(double));
+    double * vector2 = (double*)malloc(s1 * sizeof(double));
+    double** matrix1 = (double**)malloc(s1 * sizeof(double*));
     for (int i = 0; i < s1; i++) {
-        matrix1[i] = (float*)malloc(s2 * sizeof(float));
+        matrix1[i] = (double*)malloc(s2 * sizeof(double));
     }
 
     for (int i = 0; i < s1; i++) {
@@ -227,7 +227,7 @@ void testJacoby() {
     }
 
     double time;
-    float* result = jacoby(matrix1, vector2, vector1, 0.001, 1000, s1, &time);
+    double* result = jacoby(matrix1, vector2, vector1, 0.001, 1000, s1, &time);
 
     printf("Results:\n");
     for (int i = 0; i < s1; i++) {
